+++ app/routes/components/Graph/changelog.codegen.ts
@@ -5,4 +5,6 @@ export const changelog = ''
   + '- 2025-11-16: Initial version.\n'
   + '- 2025-11-23: Refactor the code to use constants. Let the edges use bezier\n'
   + '  curves. Support dark mode.\n'
+  + '- 2025-12-07: Use Canvas API to measure text width to handle the node with long\n'
+  + '  label.\n'
   ;