+++ src/components/Graph/index.ts
@@ -1,3 +1,3 @@
-// Graph from copy-ui @ 2025-11-23.
+// Graph from copy-ui @ 2025-12-07.
 
 export { Graph } from "./graph";