+++ src/components/Graph/CHANGELOG.md
@@ -1,3 +1,5 @@
 - 2025-11-16: Initial version.
 - 2025-11-23: Refactor the code to use constants. Let the edges use bezier
   curves. Support dark mode.
+- 2025-12-07: Use Canvas API to measure text width to handle the node with long
+  label.